Consolidate check utf8
